improve docstrings & logging; correct pep8 violations; let text utilities be accessible to etl.errors
