+++ python/etl/data_warehouse.py
@@ -24,7 +24,7 @@ import etl.config.dw
 import etl.db
 from etl.config.dw import DataWarehouseSchema
 from etl.errors import ETLRuntimeError, ETLConfigError
-from etl.names import join_with_quotes
+from etl.text import join_with_quotes
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
