+++ python/etl/text.py
@@ -1,8 +1,37 @@
 """
-Deal with text, mostly around matrices of texts which we want to pretty print.
+Deal with text, mostly around iterables of texts which we want to pretty print.
+Should not import Arthur modules (so that etl.errors remains widely importable)
 """
 
 import textwrap
+
+
+def join_with_quotes(names):
+    """
+    Individually wrap the names in quotes and return comma-separated names in a string.
+
+    If the input is a set of names, the names are sorted first.
+    If the input is a list of names, the order of the list is respected.
+    If the input is cheese, the order is for more red wine.
+
+    >>> join_with_quotes(["foo", "bar"])
+    "'foo', 'bar'"
+    >>> join_with_quotes({"foo", "bar"})
+    "'bar', 'foo'"
+    >>> join_with_quotes(frozenset(["foo", "bar"]))
+    "'bar', 'foo'"
+    """
+    if isinstance(names, (set, frozenset)):
+        return ', '.join("'{}'".format(name) for name in sorted(names))
+    else:
+        return ', '.join("'{}'".format(name) for name in names)
+
+
+def join_column_list(columns):
+    """
+    Return string with comma-separated, delimited column names
+    """
+    return ", ".join('"{}"'.format(column) for column in columns)
 
 
 class ColumnWrapper(textwrap.TextWrapper):
